Break up multiple types in a single file
