+++ src/binstarjs03.AerialOBJ.WpfApp/Settings/PerformancePreference.cs
@@ -1,0 +1,7 @@
+﻿namespace binstarjs03.AerialOBJ.WpfApp;
+
+public enum PerformancePreference
+{
+    OptimalMemoryUsage,
+    OptimalPerformance,
+}
